requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.8
matplotlib>=3.10
seaborn>=0.13
streamlit>=1.55.0
altair>=5.0
joblib>=1.5
shap>=0.49
xgboost>=3.2
tensorflow-cpu==2.20.0; python_version < "3.13"
